Implement KSQL function type checks

+++ tests/Query/Builders/HavingClauseBuilderTests.cs
@@ -29,8 +29,8 @@ public class HavingClauseBuilderTests
 
     private class Order
     {
-        public decimal Amount { get; set; }
-        public decimal Price { get; set; }
+        public double Amount { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; }
     }
